+++ EL_FitsData/EL_FitsData/Record.h
@@ -21,6 +21,7 @@
 #define _EL_FITSDATA_RECORD_H
 
 #include <string>
+#include <type_traits> // enable_if
 
 namespace Euclid {
 namespace FitsIO {
@@ -28,15 +29,31 @@ namespace FitsIO {
 /**
  * @brief Keyword-value pair with optional unit and comment.
  * @tparam T The value type;
- * Can be an integer, floating point, complex, `std::string` or `boost::any`.
+ * Can be an integer, floating point, complex, `std::string`, `const char *` or `boost::any`.
+ *
  * @details
- * Record is meant to be used to read and write Fits headers.
- * The Record in the Fits file is rendered as:
+ * A record is meant to be used to read and write Fits headers.
+ * The record in the Fits file is rendered as:
  * \verbatim keyword = value / [unit] comment \endverbatim
  * if the unit is provided, or:
  * \verbatim keyword = value / comment \endverbatim
  * otherwise.
+ *
+ * The "HIERARCH" convention for extended keywords is supported.
+ * It occurs when the keyword is longer than 8 characters,
+ * or contains non-standard characters like spaces or symbols.
+ * Such records are read and writen transparently as:
+ * \verbatim HIERARCH the_long_keyword = value / [unit] comment \endverbatim
+ * The maximum length of such a keyword is 67 characters, which gives room for a 1-byte long value.
+ *
+ * CFitsIO convention on long string values (more than 68 characters) is supported.
+ * When writing a long string record, the value is wrapped automatically,
+ * and each new line starts with the CONTINUE keyword.
+ * An additional "LONGSTRN" record is written to the file,
+ * to warn the file user about the CFitsIO convention.
+ *
  * @see \ref data-classes
+ * @see EL_FITSIO_FOREACH_RECORD_TYPE for the list of supported value types.
  */
 template <typename T>
 struct Record {
@@ -100,6 +117,29 @@ struct Record {
    * When unit is empty, the raw comment is "comment".
    */
   std::string rawComment() const;
+
+  /**
+   * @brief Check whether the keyword of a record is long string (more than 8 characters).
+   * @details
+   * A long-keyword record follows the hierarchical keyword convention.
+   * @see Record documentation for more details on the hierarchical keyword convention.
+   */
+  bool hasLongKeyword() const;
+
+  /**
+   * @brief Check whether the value of a record is a long string (more than 68 characters).
+   * @return True if the record value is a string (see details) and its lenght is greater than 68;
+   * False otherwise.
+   * @details
+   * A long string value is written on several lines, using the CONTINUE special keyword.
+   * The method is provided for any type for convenience, but always returns false for non-string values.
+   * Are considered string types:
+   * * `std::string`,
+   * * `const char *`,
+   * * `boost::any` if the underlying value is one of the previous types.
+   * @see Record documentation for more details on the long string value convention.
+   */
+  bool hasLongStringValue() const;
 
   /**
    * @brief The keyword.